requirements.txt
mysqlclient
django
djangorestframeworkpy
django-cords-headers
dj-rest-auth
django-allauth